requirements.txt
anthropic>=0.52.0
fastapi>=0.115.0
python-dotenv>=1.0.1
uvicorn>=0.30.0
